+++ src/schema.rs
@@ -1,13 +1,18 @@
 //! Logic for parsing and interacting with schemas in Avro format.
 use std::borrow::Cow;
 use std::collections::HashMap;
+use std::fmt;
 
-use failure::Error;
-use serde::ser::{Serialize, SerializeMap, SerializeSeq, Serializer};
+use digest::Digest;
+use failure::{Error, Fail};
+use serde::{
+    ser::{SerializeMap, SerializeSeq},
+    Serialize, Serializer,
+};
 use serde_json::{self, Map, Value};
 
-use types;
-use util::MapHelper;
+use crate::types;
+use crate::util::MapHelper;
 
 /// Describes errors happened while parsing Avro schemas.
 #[derive(Fail, Debug)]
@@ -20,6 +25,27 @@ impl ParseSchemaError {
         S: Into<String>,
     {
         ParseSchemaError(msg.into())
+    }
+}
+
+/// Represents an Avro schema fingerprint
+/// More information about Avro schema fingerprints can be found in the
+/// [Avro Schema Fingerprint documentation](https://avro.apache.org/docs/current/spec.html#schema_fingerprints)
+pub struct SchemaFingerprint {
+    pub bytes: Vec<u8>,
+}
+
+impl fmt::Display for SchemaFingerprint {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        write!(
+            f,
+            "{}",
+            self.bytes
+                .iter()
+                .map(|byte| format!("{:02x}", byte))
+                .collect::<Vec<String>>()
+                .join("")
+        )
     }
 }
 
@@ -335,7 +361,7 @@ impl UnionSchema {
 
     /// Optionally returns a reference to the schema matched by this value, as well as its position
     /// within this enum.
-    pub fn find_schema(&self, value: &::types::Value) -> Option<(usize, &Schema)> {
+    pub fn find_schema(&self, value: &crate::types::Value) -> Option<(usize, &Schema)> {
         let kind = SchemaKind::from(value);
         self.variant_index
             .get(&kind)
@@ -376,6 +402,20 @@ impl Schema {
     pub fn canonical_form(&self) -> String {
         let json = serde_json::to_value(self).unwrap();
         parsing_canonical_form(&json)
+    }
+
+    /// Generate [fingerprint] of Schema's [Parsing Canonical Form].
+    ///
+    /// [Parsing Canonical Form]:
+    /// https://avro.apache.org/docs/1.8.2/spec.html#Parsing+Canonical+Form+for+Schemas
+    /// [fingerprint]:
+    /// https://avro.apache.org/docs/current/spec.html#schema_fingerprints
+    pub fn fingerprint<D: Digest>(&self) -> SchemaFingerprint {
+        let mut d = D::new();
+        d.input(self.canonical_form());
+        SchemaFingerprint {
+            bytes: d.result().to_vec(),
+        }
     }
 
     /// Parse a `serde_json::Value` representing a primitive Avro type into a
@@ -801,7 +841,8 @@ mod tests {
                 ]
             }
         "#,
-        ).unwrap();
+        )
+        .unwrap();
 
         let mut lookup = HashMap::new();
         lookup.insert("a".to_owned(), 0);
@@ -912,4 +953,33 @@ mod tests {
         sync(&schema);
         sync(schema);
     }
+
+    #[test]
+    fn test_schema_fingerprint() {
+        use md5::Md5;
+        use sha2::Sha256;
+
+        let raw_schema = r#"
+    {
+        "type": "record",
+        "name": "test",
+        "fields": [
+            {"name": "a", "type": "long", "default": 42},
+            {"name": "b", "type": "string"}
+        ]
+    }
+"#;
+
+        let schema = Schema::parse_str(raw_schema).unwrap();
+        assert_eq!(
+            "c4d97949770866dec733ae7afa3046757e901d0cfea32eb92a8faeadcc4de153",
+            format!("{}", schema.fingerprint::<Sha256>())
+        );
+
+        assert_eq!(
+            "7bce8188f28e66480a45ffbdc3615b7d",
+            format!("{}", schema.fingerprint::<Md5>())
+        );
+    }
+
 }